+++ attributes/default.rb
@@ -3,6 +3,9 @@
 # Attributes:: default
 #
 
+# if you still need deprecated recipes related to tuning Solr and Postgres, set this back to 13 or 12
+default['mcs']['chef_server_version'] = 14
+
 # upgrade package
 default['mcs']['upgrade']['package_source'] = nil
 
@@ -65,9 +68,6 @@ default['mcs']['policyfile']['dir'] = nil
 default['mcs']['policyfile']['group'] = '_default'
 default['mcs']['policyfile']['lockfiletype'] = '.lock.json'
 default['mcs']['policyfile']['purge'] = false
-
-# added to allow skipping chef-server-ctl test for already deployed systems
-default['mcs']['skip_test'] = false
 
 # _tuning recipe
 # if you want to configure the settings in the _tuning recipe you may set these