+++ test/integration/default/default_test.rb
@@ -50,8 +50,8 @@ end
 
 describe file '/etc/opscode/chef-server.rb' do
   it { should exist }
-  its('content') { should match /^opscode_solr4\['heap_size'\] = / }
-  its('content') { should match /^postgresql\['checkpoint_completion_target'\] = 0.9$/ }
-  its('content') { should match /^postgresql\['checkpoint_segments'\] = 64$/ }
-  its('content') { should match /^postgresql\['log_min_duration_statement'\] = 1000$/ }
+  # its('content') { should match /^opscode_solr4\['heap_size'\] = / }
+  # its('content') { should match /^postgresql\['checkpoint_completion_target'\] = 0.9$/ }
+  # its('content') { should match /^postgresql\['checkpoint_segments'\] = 64$/ }
+  # its('content') { should match /^postgresql\['log_min_duration_statement'\] = 1000$/ }
 end